refactor(tasks): fix imports
