+++ contracts/dollar/tasks/library/price-reset/dryRunner.ts
@@ -1,5 +1,5 @@
 import { Signer } from "ethers";
-import { TaskArgs } from ".";
+import { TaskArgs } from '../price-reset';
 
 const ADMIN_WALLET = "0xefC0e701A824943b469a694aC564Aa1efF7Ab7dd";
